Use new RIDE logo and simplify splash screen
This changes the background of the splash screen to be the normal dialog background colour (rather than purple) and uses the purple for the logotype, which is also used in the About dialog. Throughout, RIDE is used rather than Dyalog. Notably, these changes speed up the launch process so significantly that the splash screen is hardly shown, which is another reason it should match background colour with the connect screen, preventing a colour change flash.